update supporting scripts and makefile

+++ support/build_vim-tools.sh
@@ -15,19 +15,24 @@ die()
 }
 
 # get source code
-echo "getting sources ..."
-git clone "${REPO}" "${DEST}" || die "get sources failed"
+if [ ! -s ${DEST} ]
+then
+  echo -n "getting vim-tools ... "
+  git clone --depth=1 "${REPO}" "${DEST}" >/dev/null 2>&1 || die "FAILED"
+  echo "DONE"
+fi
 
 # virtualenv
-echo "creating virtualenv ..."
-virtualenv ${DEST}/virtualenv || die "virtualenv failed"
-${DEST}/virtualenv/bin/pip install -r ${DEST}/requirements.txt || die "virtualenv failed"
+echo -n "creating virtualenv ... "
+virtualenv ${DEST}/virtualenv >/dev/null 2>&1 || die "FAILED"
+${DEST}/virtualenv/bin/pip install -r ${DEST}/requirements.txt >/dev/null 2>&1 || die "FAILED"
+echo "DONE"
 
 # executable
 mkdir ${DEST}/bin
 cat > ${DEST}/bin/html2vimdoc <<EOF
 #!/bin/sh
-. ${DEST}/virtualenv/bin/activate
+. ${DEST}/virtualenv/bin/activate >/dev/null 2>&1
 ${DEST}/html2vimdoc.py \$*
 EOF
 chmod 0755 ${DEST}/bin/html2vimdoc
